# hek_analysis/docs/architecture/Unified-Healthcare-API_openapi.yaml
openapi: 3.0.3
info:
  title: Unified Healthcare API
  description: >
    Canonical contract consolidating HISO, KARO, and ERMS. See
    Unified-Healthcare-API_API-Contract-Design.md for full rationale,
    traceability, and reconciliation decisions. No version identifier
    appears in the URL/header/media-type per stakeholder decision (Section 9
    of the contract document) — version is tracked here via info.version and
    an append-only changelog only; no request/response shape changes without
    a coordinated migration.
  version: "1.1.1"
  x-changelog:
    - version: "1.0.0"
      date: "2026-07-19"
      description: Initial canonical contract, Phase 7 (API Contract Design).
    - version: "1.1.0"
      date: "2026-07-19"
      description: >
        Demographics split into 4 legacy-shaped endpoints instead of one merged
        endpoint; HISO's addInvoice/launchForm actions and static-mode form
        retrieval implemented (previously excluded as dead stubs); KARO's
        GetEncounterSummary and SaveScreeningCode implemented for real
        (previously excluded as confirmed-broken); rate limiting confirmed
        as a built, config-toggled capability.
    - version: "1.1.1"
      date: "2026-07-20"
      description: >
        Block 3 contract re-verification against the actual Day-1 build (all 42
        operations implemented, route-for-route and verb-for-verb match).
        One drift corrected here rather than left silently mismatched:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /patients/{patientId}/demographics/hiso:
    get:
      tags: [Demographics]
      summary: Retrieve HISO-shaped patient demographics (kept separate per stakeholder decision, not merged)
      parameters:
        - $ref: '#/components/parameters/patientId'
      responses:
        '200':
          description: HISO-shaped demographic data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemographicsHiso'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /patients/{patientId}/demographics/karo:
    get:
      tags: [Demographics]
      summary: Retrieve KARO-shaped patient demographics (kept separate per stakeholder decision, not merged)
      parameters:
        - $ref: '#/components/parameters/patientId'
      responses:
        '200':
          description: KARO-shaped demographic data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemographicsKaro'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /patients/{patientId}/demographics/erms:
    get:
      tags: [Demographics]
      summary: Retrieve ERMS-shaped patient demographics (kept separate per stakeholder decision, not merged)
      parameters:
        - $ref: '#/components/parameters/patientId'
      responses:
        '200':
          description: ERMS-shaped demographic data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemographicsErms'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /patients/{patientId}/demographics/col:
    get:
      tags: [Demographics]
      summary: Retrieve COL/Pegasus-shaped patient demographics (kept separate per stakeholder decision, not merged)
      parameters:
        - $ref: '#/components/parameters/patientId'
      responses:
        '200':
          description: COL/Pegasus-shaped demographic data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemographicsCol'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /patients/{patientId}/documents:
    get:
      tags: [Documents]
      summary: List documents
      parameters:
        - $ref: '#/components/parameters/patientId'
        - name: direction
          in: query
          schema: { type: string, enum: [in, out] }
        - name: contentType
          in: query
          schema: { type: string }
        - name: referenceId
          in: query
          schema: { type: string }
        - name: subject
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/sinceDate'
        - $ref: '#/components/parameters/untilDate'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of documents (metadata only, no content)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DocumentSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Documents]
      summary: Save a document
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentInput'
      responses:
        '201':
          description: Document saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/{patientId}/documents/{documentId}:
    get:
      tags: [Documents]
      summary: Retrieve document content
      parameters:
        - $ref: '#/components/parameters/patientId'
        - name: documentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Document content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /patients/{patientId}/encounters/{encounterId}/observations:
    get:
      tags: [Observations]
      summary: Retrieve observations/measurements
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - name: conceptId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of observations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Observation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Observations]
      summary: Save observations (at least one measurement field required, KARO-BR-14)
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationInput'
      responses:
        '201':
          description: Observation saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Observation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /encounter-summary-templates/{identifier}/schema:
    get:
      tags: [Encounter Summary]
      summary: Schema-driven field/caption/type lookup for a template (KARO-BR-10)
      parameters:
        - name: identifier
          in: path
          required: true
          schema: { type: string }
          description: e.g. diap, cvra, foot-exam, retinopathy
      responses:
        '200':
          description: Template schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier: { type: string }
                  fields:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        caption: { type: string }
                        type: { type: string }
  /patients/{patientId}/encounters/{encounterId}/encounter-summary:
    get:
      tags: [Encounter Summary]
      summary: >
        Retrieve a patient's saved encounter summary. Legacy GetEncounterSummary
        returned hardcoded mock data unrelated to the requested patient
        (KARO-BR-09); implemented for real per stakeholder decision 2026-07-19.
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - name: identifier
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Saved encounter summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier: { type: string }
                  fields:
                    type: object
                    additionalProperties: true
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Encounter Summary]
      summary: Save a templated encounter summary (typed validation against schema, replacing loose JObject parsing)
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, fields]
              properties:
                identifier: { type: string }
                fields:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Encounter summary saved
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /recall-categories:
    get:
      tags: [Recalls]
      summary: List recall categories
      parameters:
        - name: group
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Recall categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        categoryId: { type: string }
                        name: { type: string }
  /patients/{patientId}/recalls:
    get:
      tags: [Recalls]
      summary: List recalls for a patient
      parameters:
        - $ref: '#/components/parameters/patientId'
      responses:
        '200':
          description: List of recalls
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Recall' }
    post:
      tags: [Recalls]
      summary: Save a recall (empty categoryId defaults per group, KARO-BR-22)
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecallInput'
      responses:
        '201':
          description: Recall saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recall' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Resource-scoped token per ADR-003 (patientId + encounterId + practiceId +
        structurally-determined originScope). Issued via POST /auth/token, or
        internally minted by the HISO SOAP edge adapter from an existing
        SessionGUID (ADR-004/ADR-007).
  parameters:
    patientId:
      name: patientId
      in: path
      required: true
      schema: { type: integer }
    encounterId:
      name: encounterId
      in: path
      required: true
      schema: { type: integer }
...
    sinceDate:
      name: sinceDate
      in: query
      schema: { type: string, format: date }
    untilDate:
      name: untilDate
      in: query
      schema: { type: string, format: date }
    sortOrder:
      name: sortOrder
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: See Section 12 of the contract document. Optional; falls back to natural-key duplicate detection if omitted.
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Valid credential, insufficient scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      description: Data-minimized error shape — never contains PHI, stack traces, or internal exception detail (SRS §12.14).
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: Generic, static, non-PHI message. }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
      required: [title, status]
...
    DemographicsHiso:
      type: object
      description: HISO's concept-driven demographics shape. Kept separate, not merged (Section 8 Decision 6 of the contract doc).
      properties:
        patientId: { type: integer }
        practiceId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string, format: date }
    DemographicsKaro:
      type: object
      description: KARO's DemographicInfo shape, per KARO_HSS_doc.md sample. Kept separate, not merged.
      properties:
        patientId: { type: integer }
        practiceId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string, format: date }
        dateOfEnrolment: { type: string, format: date, nullable: true }
        endEnrolmentDate: { type: string, format: date, nullable: true }
    DemographicsErms:
      type: object
      description: ERMS APIController's PatientData shape (JSON-translated from legacy XML). Kept separate, not merged.
      properties:
        patientId: { type: integer }
        encounterId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        dob: { type: string, format: date }
        nhi: { type: string, nullable: true }
    DemographicsCol:
      type: object
      description: COL/Pegasus PegasusAPIModel.PatientData shape. Kept separate, not merged.
      properties:
        patientId: { type: integer }
        practiceId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
...
    DocumentSummary:
      type: object
      properties:
        documentId: { type: string }
        patientId: { type: integer }
        direction: { type: string, enum: [in, out] }
        contentType: { type: string }
        createdAt: { type: string, format: date-time }
        subject: { type: string, nullable: true }
        referenceId: { type: string, nullable: true }
    Document:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          properties:
            content: { type: string, format: byte, description: Base64-encoded document content }
    DocumentInput:
      type: object
      required: [direction, contentType, content]
      properties:
        direction: { type: string, enum: [in, out] }
        contentType: { type: string }
        subject: { type: string }
        referenceId: { type: string }
        content: { type: string, format: byte }
    Observation:
      type: object
      properties:
        observationId: { type: string }
        conceptId: { type: string }
        value: { type: string }
        recordedAt: { type: string, format: date-time }
    ObservationInput:
      type: object
      description: At least one of the nine measurement fields must be non-empty (KARO-BR-14) — enforced at the API layer, not just documented.
      properties:
        height: { type: number, nullable: true }
        weight: { type: number, nullable: true }
        bmi: { type: number, nullable: true }
        bloodPressureSystolic: { type: number, nullable: true }
        bloodPressureDiastolic: { type: number, nullable: true }
        waistCircumference: { type: number, nullable: true }
        smokingStatus: { type: string, nullable: true }
        heartRate: { type: number, nullable: true }
        temperature: { type: number, nullable: true }
...
    Recall:
      type: object
      properties:
        recallId: { type: string }
        patientId: { type: integer }
        categoryId: { type: string, nullable: true }
        dueDate: { type: string, format: date }
    RecallInput:
      type: object
      properties:
        categoryId: { type: string, nullable: true, description: If empty, defaults to the first category for the group at the stored-procedure level (KARO-BR-22). }
        dueDate: { type: string, format: date }
...